4) Trate as exceções na classe de integração com ViaCEP
