+++ src/main/java/br/com/alura/comex/ClienteTeste.java
@@ -3,8 +3,7 @@ package br.com.alura.comex;
 public class ClienteTeste {
 
     public static void main(String[] args) {
-        Endereco endereco = new Endereco("Jardim das Flores", "São Paulo", "Apartamento 101", "SP", "Rua das Camélias",
-                123);
+        Endereco endereco = new Endereco("71680349");
 
         Cliente cliente = new Cliente("João da Silva", "123.456.789-00", "joao.silva@example.com", "Engenheiro",
                 "11 99999-9999", endereco);
