+++ src/main/java/br/com/alura/comex/TestaIntegracaoViaCep.java
@@ -5,7 +5,7 @@ import java.io.IOException;
 public class TestaIntegracaoViaCep {
     public static void main(String[] args) throws IOException, InterruptedException {
         try {
-            System.out.println(APIViaJava.getAddress("71680349"));
+            System.out.println(APIViaCep.getAddress("71680349"));
         } catch (RuntimeException e) {
             System.out.println(e.getMessage());
         }
